requirements.txt
boto3==1.9.203
botocore==1.12.203
Click==7.0
docutils==0.14
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
pygame==1.9.6
python-dateutil==2.8.0
python-redis-lock==3.3.1
redis==3.3.6
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5

